Created Collision Layers
Created separate collision layers for the player, grabbable, and the floor.  Then set it so that the player cannot interact with grabbables.  All grabbable objects were placed in the grabbables layer, and the player body was placed in the player layer.  The player hands though were left in the default layer.